Split survey and stats opt-in handlers out of PlayerTitleHelper
PlayerTitleHelper had accumulated three unrelated concerns: player
titles, the start-of-game preferences survey and its table-settings
buttons, and the end-of-game statistics opt-in.

Move the survey and table-settings handlers (answerSurvey,
purgeSupports, purgeOverrule, noSupportSwaps, setLimitedWhispers) to a
new GameSurveyService, alongside the StartPhaseService code that posts
those buttons. Move the opt-in buttons (getOptInButtons,
setOptInSinglePreference, setOptInStats) into StatisticsOptInOutService,
which already held the slash command half of that feature.

Method bodies are unchanged; AnnotationHandler scans all of ti4 for
@ButtonHandler, so no registration changes are needed.
